migrate build pipeline to esbuild for ESM-only @actions/* support
Replace @vercel/ncc with esbuild, which can bundle ESM-only packages into
a CJS output file. Upgrade @actions/cache, core, http-client, and tool-cache
to their latest versions (which are ESM-only). Update tsconfig to use
moduleResolution: "bundler" so TypeScript can resolve ESM package exports
during type-checking.

Split pure utility functions into src/flutter.ts (no external deps) so that
Jest can continue running in CJS mode via a separate tsconfig.test.json
without ever importing the ESM-only @actions/* packages.
