+++ package.json
@@ -2,17 +2,16 @@
   "name": "setup-flutter-action",
   "version": "1.0.0",
   "description": "A GitHub action to install the Flutter SDK.",
-  "main": "dist/index.js",
   "scripts": {
-    "build": "tsc && ncc build lib/index.js -o dist --minify",
+    "build": "tsc --noEmit && esbuild src/index.ts --bundle --platform=node --format=cjs --outfile=dist/index.js --minify",
     "lint": "eslint src/",
     "test": "jest"
   },
   "dependencies": {
-    "@actions/cache": "^5.0.5",
-    "@actions/core": "^2.0.3",
-    "@actions/http-client": "^3.0.2",
-    "@actions/tool-cache": "^3.0.1"
+    "@actions/cache": "^6.0.0",
+    "@actions/core": "^3.0.0",
+    "@actions/http-client": "^4.0.0",
+    "@actions/tool-cache": "^4.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^10.0.1",
@@ -20,7 +19,7 @@
     "@types/node": "^25.3.3",
     "@typescript-eslint/eslint-plugin": "^8.56.1",
     "@typescript-eslint/parser": "^8.56.1",
-    "@vercel/ncc": "^0.38.1",
+    "esbuild": "^0.27.3",
     "eslint": "^10.0.2",
     "globals": "^17.4.0",
     "jest": "^30.2.0",
@@ -28,7 +27,9 @@
     "typescript": "^5.5.2"
   },
   "jest": {
-    "preset": "ts-jest",
+    "transform": {
+      "^.+\\.ts$": ["ts-jest", {"tsconfig": "tsconfig.test.json"}]
+    },
     "testEnvironment": "node",
     "testMatch": [
       "**/src/**/*.test.ts"
